Update patch set 1
Patch Set 1: Verified; Looks good to me, but someone else must approve

I've ported the bugfixes from CAF without all of the 50001 voodoo that we can't do anything with due to proprietary headers.

I've tested this on Vision, Supersonic, and Passion, all with great results. I've never seen an N1 get a lock this fast, and all the weird states that I've been able to put my GPS into seem to be resolved.

You'll want to set BOARD_USES_QCOM_LIBRPC := true in your BoardConfig before testing.

I regenerated the stubs for AMSS 6225 and 3200 so Z71 and Bravo are the ones that needs tested the most, although Bravo has the same hardware as Passion so I don't expect issues.

Patch-set: 1