Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Works fine on z71.
I'm getting an RPC error when injecting a location, but after rolling back I realized it was already there with the previous version, so it's something that needs to be fixed in 6225. As far as this particular commit goes, +1.

Patch-set: 1