Update patch set 3
Patch Set 3: Code-Review-1

I'm going to go out and block this because we need to stop evaluating legacy support items on a patch by patch basis. (regardless of vestigial code bits still remaining in the CM codebase) It makes our position unclear. That is basically the gist of the previous discussion.

It looks like this was meant to support 7x30 to begin with, which is an unsupported legacy device , and thus I'm going to -1.

Patch-set: 3
Label: Code-Review=-1