Update patch set 3
Patch Set 3: Verified+1 Code-Review+1

Given the minimality of this change (it changes no code, and only alters the behavior of the build process for currently unsupported hardware, which may eventually be supported again), I don't see any reason to outright block it.  In fact, I'm going to go ahead and +1 it. :P

I can see the merit in rejecting hacks for unsupported hardware, but this is not a hack, and is minimally intrusive.

Patch-set: 3
Reviewer: Christopher Lais <119@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1
Label: Code-Review=+1