Update patch set 3
Patch Set 3: Code-Review+1

Whatever the reason or whatever device this fixes surely is not important on a code review. There is code in the repository already relating to AMSS 1240 support that is broken and this fixes it. Having a blanket ban on merging unobtrusive fixes because it just happens to be used on a "legacy" device at the moment seems short sighted. Fix it or get rid of it entirely.

Patch-set: 3
Label: Code-Review=+1