Update patch set 3
Patch Set 3:

Ok, so this workaround should be good enough then to fix the rollover for those currently affected, given that we can't exactly change the hardware. 8960, 8974 and 8916 are confirmed to have the issue. I don't know anyone else with the other chipsets to ask, but it would be no suprise they suffer the same.

Of course, users using these HALs can fork it to their device tree and implement the fix there, but that's not exactly a desired solution.

Patch-set: 3
Reviewer: Gerrit User 444 <444@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>